moved instrumentor objects to object package
